Improving EXPORT method: images -> ASCII files
In the previous version when exporting scanned images to ASCII files if
the image has size bigger than 256x256 pixels the software hangs. The
reason for the hanging (slow conversion) was the string operations in
the underlying for-loops. In the proposed commit the string operation
are avoided and thus the speed of conversion accelerated (basically
instant conversion).

Additionally (by still commented) a prototype of image conversion is
added. Needs some further improvement to work
in a better way - a pixel value may be cut if it is too higher
(otherwise the conversion works OK).